pyproject.toml
[tool.poetry]
name = "target-google-people"
version = "0.0.1"
description = "`target-google-people` is a Singer target for GooglePeople, built with the Meltano SDK for Singer Targets."
authors = ["Ken Payne"]
keywords = [
    "ELT",
    "GooglePeople",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.11,>=3.7.1"
requests = "^2.25.1"
singer-sdk = "^0.4.5"
google-api-python-client = "^2.44.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "^21.9b0"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "target_google_people"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
target-google-people = 'target_google_people.target:TargetGooglePeople.cli'
